requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
airtrain==0.1.61
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
backoff==2.2.1
black==25.1.0
boto3==1.37.28
botocore==1.37.28
CacheControl==0.14.2
cachetools==5.5.2
cerebras_cloud_sdk==1.28.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
distro==1.9.0
eval_type_backport==0.2.2
filelock==3.18.0
firebase-admin==6.7.0
fireworks-ai==0.15.12
flake8==7.2.0
frozenlist==1.5.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-genai==1.9.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
griffe==1.7.2
groq==0.22.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
huggingface-hub==0.30.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mcp==1.6.0
mdurl==0.1.2
modelcontextprotocol==0.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.3.2
mypy==1.15.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.4
openai==1.70.0
openai-agents==0.0.8
packaging==24.2
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.3.7
pluggy==1.5.0
posthog==3.23.0
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers @ git+https://github.com/UKPLab/sentence-transformers.git@8d73d4f0fceee03ce68df1c8e7f0c0fe843c4988
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.6.0
together==1.5.5
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.0
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3==2.3.0
uv==0.6.12
uvicorn==0.34.0
wcwidth==0.2.13
websockets==15.0.1
win32_setctime==1.2.0
wsproto==1.2.0
yarl==1.19.0
